Staff upgraddes

+++ staff/dave.ino
@@ -2,15 +2,15 @@
 // A gradient background, with rotating bright points and white sparkles.
 
 namespace Dave {
-  uint8_t daveOffset = 0;
+  uint16_t daveOffset = 0;
   int daveMilliseconds = 0;
   
   void draw() {
     for (int k=0; k<NUM_LEDS-1; k++) {
-      uint8_t k8 = k;
-      uint8_t hue = (k/15 * 255) + (daveOffset*5);
-      uint8_t saturation = rand() % 255 > 252 ? 0 : 255; //NUM_LEDS - (k+daveOffset*10);
-      uint8_t value = k8 % 14 == daveOffset % 14 ? 200 : 40; //NUM_LEDS- (k+daveOffset*10);
+      uint16_t k8 = k;
+      uint16_t hue = (k/15 * 255) + (daveOffset*5);
+      uint16_t saturation = rand() % 255 > 252 ? 0 : 255; //NUM_LEDS - (k+daveOffset*10);
+      uint16_t value = k8 % 14 == daveOffset % 14 ? 200 : 40; //NUM_LEDS- (k+daveOffset*10);
       if(saturation == 0){
         value = 255;
       }
